Move contents of linter class from bin directory to lib
